refactor(direct-control): extract capability catalog options
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves Civ7CapabilityCatalogOptions into src/catalog/capabilities.ts while keeping public facade type re-exports in index.ts. Removes the catalog module type-only dependency on facade root-inspection exports by importing root inspection types from their runtime owner.

Preserves catalog option shape and does not change catalog source, scanner, schema, sorting, runtime inspection, or wrapper behavior. Remaining facade-owned lifecycle/health result types, public procedure schemas, operation/proof telemetry, hotseat runtime proof, AI ingestion, CLI semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance remain pending.

Validation:
- bun run --cwd packages/civ7-direct-control test -- runtime-and-catalog.test.ts public-api.test.ts
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- git diff --cached --check
- ownership/no-cycle scan for catalog options and facade imports

Local package/type relocation proof only; no runtime, live-game, AI-on-hotseat, telemetry, semantic CLI, or procedure-core proof claimed.
