+++ openspec/changes/civ7-support-direct-control-modularization/workstream/workstream-record.md
@@ -1234,6 +1234,18 @@ All future agent waves must be framed before delegation:
   AI ingestion, CLI semantic projection, Effect/oRPC procedure-core work, and
   Task 2.9.4 matrix-row acceptance remain pending. This is local package/type
   relocation proof only, not runtime/live-game proof.
+- Direct-control capability catalog options type slice: completed as a narrow
+  public type ownership relocation. It moves `Civ7CapabilityCatalogOptions`
+  into `src/catalog/capabilities.ts` while keeping public facade type
+  re-exports in `index.ts` and removing the catalog module's type-only
+  dependency on facade root-inspection exports in favor of the runtime owner.
+  This preserves catalog option shape and does not change catalog source,
+  scanner, schema, sorting, runtime inspection, or wrapper behavior. Remaining
+  facade-owned lifecycle/health result types, public procedure schemas,
+  operation/proof telemetry, hotseat runtime proof, AI ingestion, CLI semantic
+  projection, Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row
+  acceptance remain pending. This is local package/type relocation proof only,
+  not runtime/live-game proof.
 - Direct-control diplomacy closeout source slice: completed as a narrow
   operation source relocation. It moves only the App UI diplomacy response
   closeout command builder and embedded closeout source into
